added support for seperate startIdleFrames and endIdleFrames; now uses seperate nameSv and nameEn depending on locale; removed scrolling between levels; added support for animation curves for sprites and the thought bubble; minor improvements

+++ lib/screens/games/enclose_moose/enclose_grid.dart
@@ -242,8 +242,8 @@ class EncloseGrid extends ChangeNotifier {
 
       tile.update(
         newIsEnclosed: enclosure.contains(tile),
-        newWaitFrames: expandDistance == null ? null : expandDistance,
-        newReverseWaitFrames: distance == null ? null : (maxDistance - distance)
+        newWaitFrames: expandDistance == null ? null : expandDistance.toDouble(),  // Could multiply these to change spread speed
+        newReverseWaitFrames: distance == null ? null : (maxDistance - distance).toDouble()
       );
     }
